fix(shelly): review fix - extracting types to a separate file
