+++ extensions/shelly/lib/getTrackData/getTrackData.ts
@@ -1,225 +1,18 @@
 import { type AwellSdk } from '@awell-health/awell-sdk'
 import { isNil, isEmpty, get } from 'lodash'
-
-interface GetTrackDataInput {
-  awellSdk: AwellSdk
-  pathwayId: string
-  trackId: string
-  currentActivityId: string
-}
-
-// simplified interfaces for our query responses since we will be cherry-picking the data we need
-interface PathwayTrack {
-  id: string
-  title?: string
-}
-
-interface PathwayResponse {
-  pathway?: {
-    tracks?: PathwayTrack[]
-  }
-}
-
-interface ElementResponse {
-  id: string
-  name: string
-  label?: {
-    text?: string
-  }
-  start_date?: string
-  end_date?: string
-  status: string
-  type: string
-  stakeholders?: Array<{
-    name?: string
-  }>
-  context?: {
-    step_id?: string
-    track_id?: string
-  }
-}
-
-interface ElementsResponse {
-  elements?: ElementResponse[]
-}
-
-interface ActivityResponse {
-  id: string
-  date: string
-  action: string
-  status: string
-  resolution: string
-  subject: {
-    type: string
-    name: string
-  }
-  object: {
-    type: string
-    name: string
-    id?: string
-  }
-  indirect_object?: {
-    type?: string
-    name?: string
-  }
-  form?: {
-    id?: string
-    title?: string
-  }
-  track?: {
-    id?: string
-    title?: string
-  }
-  context?: {
-    track_id?: string
-    step_id?: string
-  }
-}
-
-interface ActivitiesResponse {
-  activities?: ActivityResponse[]
-}
-
-
-interface FormResponseAnswer {
-  question_id: string
-  value?: string
-  label?: string | null
-  value_type?: string
-}
-
-interface FormResponseData {
-  activity_id: string
-  form_id: string
-  answers: FormResponseAnswer[]
-}
-
-interface FormQuestionOption {
-  label: string
-  value: number | string
-  value_string: string
-}
-
-interface FormQuestion {
-  id: string
-  key: string
-  title: string
-  userQuestionType: string
-  options?: FormQuestionOption[]
-}
-
-interface FormDefinition {
-  id: string
-  title: string
-  key: string
-  definition_id: string
-  release_id: string
-  questions: FormQuestion[]
-}
-
-interface FormDefinitionResponse {
-  form?: {
-    form: FormDefinition
-  }
-}
-
-
-// Define data points interface - extend from Awell's DataPoint type
-interface ExtendedDataPoint {
-  id: string
-  data_set_id: string
-  key: string
-  serialized_value: string
-  data_point_definition_id: string
-  date: string
-  valueType: string
-  activity_id: string
-  definitionKey?: string
-  definitionTitle?: string
-}
-
-interface DataPointsResponse {
-  dataPoints?: ExtendedDataPoint[]
-}
-
-// Define the combined query response type
-interface CombinedQueryResponse {
-  pathway?: PathwayResponse
-  pathwayElements?: ElementsResponse
-  pathwayActivities?: ActivitiesResponse
-  pathwayDataPoints?: DataPointsResponse
-}
-
-// Type for form responses map
-type FormResponsesMap = Record<string, Record<string, FormResponseData>>;
-
-// Add a new interface for MessageResponse
-interface MessageResponse {
-  message?: {
-    message?: {
-      id: string
-      subject?: string
-      body?: string
-      format?: string
-      attachments?: Array<{
-        id: string
-        name: string
-        type: string
-        url: string
-      }> | null
-    }
-  }
-}
-
-// Simplified activity interface
-interface ExtendedActivity {
-  date: string
-  action: string
-  status: string
-  resolution?: string
-  subject: {
-    type: string
-    name: string
-  }
-  object: {
-    type: string
-    name: string
-  }
-  indirect_object?: {
-    type: string
-    name: string
-  }
-  form?: {
-    title: string
-    questions?: Array<{
-      title: string
-      response?: string
-    }>
-  }
-  message?: {
-    subject?: string
-    body?: string
-  }
-  data_points?: Array<{
-    value: string
-    title?: string
-    date: string
-  }>
-}
-
-export interface GetTrackDataOutput {
-  steps: Array<{
-    name: string
-    label: string
-    status: string
-    start_date?: string
-    end_date?: string
-    stakeholders?: Array<{
-      name: string
-    }>
-    activities: ExtendedActivity[]
-  }>
-}
+import type {
+  GetTrackDataInput,
+  GetTrackDataOutput,
+  ActivityResponse,
+  ElementResponse,
+  ExtendedDataPoint,
+  FormResponsesMap,
+  MessageResponse,
+  CombinedQueryResponse,
+  FormDefinitionResponse,
+  ExtendedActivity,
+  FormResponseAnswer
+} from './types'
 
 /**
  * Fetch and process track data from Awell SDK
